+++ test/test_rectangle.cpp
@@ -7,7 +7,7 @@ using Platec::Rectangle;
 TEST(Rectangle, MapIndexInsideRectNotWrapping)
 {
     Rectangle r = Rectangle(WorldDimension(50, 30), 42, 48, 8, 15);
-    size_t px, py, res;
+    uint32_t px, py, res;
 
     px = 42;
     py =  8;
@@ -34,7 +34,7 @@ TEST(Rectangle, MapIndexInsideRectNotWrapping)
 TEST(Rectangle, MapIndexOutsideRect)
 {
     Rectangle r = Rectangle(WorldDimension(50, 30), 42, 48, 8, 15);
-    size_t px, py, res;
+    uint32_t px, py, res;
 
     px = 49;
     py =  8;
@@ -61,7 +61,7 @@ TEST(Rectangle, MapIndexOutsideRect)
 TEST(Rectangle, MapIndexInsideRectWrappingOnX)
 {
     Rectangle r = Rectangle(WorldDimension(50, 30), 42, 6, 8, 12);
-    size_t px, py, res;
+    uint32_t px, py, res;
 
     px = 42;
     py =  8;
@@ -88,7 +88,7 @@ TEST(Rectangle, MapIndexInsideRectWrappingOnX)
 TEST(Rectangle, MapIndexInsideRectWrappingOnY)
 {
     Rectangle r = Rectangle(WorldDimension(50, 30), 42, 48, 25, 5);
-    size_t px, py, res;
+    uint32_t px, py, res;
 
     px = 42;
     py = 25;
@@ -115,7 +115,7 @@ TEST(Rectangle, MapIndexInsideRectWrappingOnY)
 TEST(Rectangle, MapIndexInsideRectLargeAsWorld)
 {
     Rectangle r = Rectangle(WorldDimension(50, 30), 0, 50, 0, 30);
-    size_t px, py, res;
+    uint32_t px, py, res;
 
     px = 0;
     py = 0;
